api/requirements.txt
flask
flask_cors
flask_sqlalchemy
flask_migrate
python-dotenv
mysqlclient
pymysql
cryptography